Create StorageFactory to embed low level code
