+++ src/app/shared/utils/string-utils.ts
@@ -1,3 +1,3 @@
 export function capitalize<T extends string>(str: T): Capitalize<T> {
-  return `${str.at(0)}${str.slice(1)}` as Capitalize<T>;
+  return `${str.charAt(0).toUpperCase()}${str.slice(1)}` as Capitalize<T>;
 }
